+++ packages/reactabular-utils/CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 
   * Bug fix - If `rowKey` value is zero, `resolveRowKey` should return `0-row`.
+  * Feature - Allow `colSpan` and `rowSpan` to be overridden at `resolveHeaderRows`.
 
 3.0.0 / 2016-09-01
 ==================